backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ahmed El-Kilani <ahmedelkilani01@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
dj-database-url = "^2.2.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
